refactor: move some components into containers directory
